rename llama-loader to ggml-loader
