+++ Cargo.toml
@@ -2,7 +2,7 @@
 members = [
     "ggml-sys",
     "ggml",
-    "llama-loader",
+    "ggml-loader",
     "llama-rs",
     "llama-cli",
     "generate-ggml-bindings"
