+++ llama-rs/src/convert.rs
@@ -17,7 +17,7 @@ use std::{
 };
 
 use crate::{util, Hyperparameters, Vocabulary};
-use llama_loader::encode_element_type;
+use ggml_loader::util::encode_element_type;
 
 /// Converts a `pth` file to a `ggml` file.
 pub fn convert_pth_to_ggml(model_directory: &Path, element_type: ggml::Type) {
